feat(entity): adding tb_main_food_ingredients refactor(entity): editing tb_main_food, tb_food_idnt

+++ src/main/java/com/bit/healthpartnerboot/entity/Member.java
@@ -23,7 +23,7 @@ public class Member {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_seq")
-    private Long seq;
+    private Integer seq;
     @Column(unique = true, nullable = false)
     private String email;
     @Column(nullable = false)
